+++ src/libxtp/gwbse/rpa.cc
@@ -20,6 +20,7 @@
 // Local VOTCA includes
 #include "votca/xtp/rpa.h"
 #include "votca/xtp/aomatrix.h"
+#include "votca/xtp/openmp_cuda.h"
 #include "votca/xtp/threecenter.h"
 #include "votca/xtp/vc2index.h"
 
@@ -58,15 +59,16 @@ template <bool imag>
 Eigen::MatrixXd RPA::calculate_epsilon(double frequency) const {
   const Index size = _Mmn.auxsize();
 
-  Eigen::MatrixXd result = Eigen::MatrixXd::Identity(size, size);
-
   const Index lumo = _homo + 1;
   const Index n_occ = lumo - _rpamin;
   const Index n_unocc = _rpamax - lumo + 1;
   const double freq2 = frequency * frequency;
   const double eta2 = _eta * _eta;
 
-#pragma omp parallel for schedule(dynamic) reduction(+ : result)
+  OpenMP_CUDA transform;
+  transform.createTemporaries(n_unocc, size);
+
+#pragma omp parallel for schedule(dynamic)
   for (Index m_level = 0; m_level < n_occ; m_level++) {
     const double qp_energy_m = _energies(m_level);
 
@@ -83,9 +85,11 @@ Eigen::MatrixXd RPA::calculate_epsilon(double frequency) const {
       sum += deltEf / (deltEf.square() + eta2);
       denom = 2 * sum;
     }
-    result += Mmn_RPA.transpose() * denom.asDiagonal() * Mmn_RPA;
-  }
 
+    transform.A_TDA(Mmn_RPA, denom);
+  }
+  Eigen::MatrixXd result = transform.A_TDA_result();
+  result.diagonal().array() += 1.0;
   return result;
 }
 
@@ -95,14 +99,16 @@ template Eigen::MatrixXd RPA::calculate_epsilon<false>(double frequency) const;
 Eigen::MatrixXd RPA::calculate_epsilon_r(std::complex<double> frequency) const {
 
   const Index size = _Mmn.auxsize();
-  Eigen::MatrixXd result = Eigen::MatrixXd::Identity(size, size);
 
   const Index lumo = _homo + 1;
   const Index n_occ = lumo - _rpamin;
   const Index n_unocc = _rpamax - lumo + 1;
+  OpenMP_CUDA transform;
+  transform.createTemporaries(n_unocc, size);
 
-#pragma omp parallel for schedule(dynamic) reduction(+ : result)
+#pragma omp parallel for schedule(dynamic)
   for (Index m_level = 0; m_level < n_occ; m_level++) {
+
     const double qp_energy_m = _energies(m_level);
     const Eigen::MatrixXd Mmn_RPA = _Mmn[m_level].bottomRows(n_unocc);
     const Eigen::ArrayXd deltaE = _energies.tail(n_unocc).array() - qp_energy_m;
@@ -116,9 +122,10 @@ Eigen::MatrixXd RPA::calculate_epsilon_r(std::complex<double> frequency) const {
     Eigen::VectorXd chi =
         deltaEm * (deltaEm.cwiseAbs2() + sigma_1).cwiseInverse() -
         deltaEp * (deltaEp.cwiseAbs2() + sigma_2).cwiseInverse();
-    result += -2 * Mmn_RPA.transpose() * chi.asDiagonal() * Mmn_RPA;
+    transform.A_TDA(Mmn_RPA, chi);
   }
-
+  Eigen::MatrixXd result = -2 * transform.A_TDA_result();
+  result.diagonal().array() += 1.0;
   return result;
 }